Refactor: 설정 페이지 컴포넌트 폴더 이동
